feat: middleware directory and resolvers
